refactor: Move `sub` implementation to a separate file
