+++ src/uni.c
@@ -1,107 +1,5 @@
 #include "uni.h"
 
-#include <grapheme.h>
-
-static inline size_t normalize_start_pos(lua_Integer pos, size_t len)
-{
-    if (pos > 0) {
-        return (size_t)pos;
-    }
-    else if (pos == 0) {
-        return 1;
-    }
-    else if (pos < -(lua_Integer)len) {
-        return 1;
-    }
-    else if (pos < 0) {
-        return (size_t)(len + pos + 1);
-    }
-    assert(!"unreachable");
-}
-
-static inline size_t normalize_end_pos(lua_Integer pos, size_t len)
-{
-    if (pos > (lua_Integer)len) {
-        return len;
-    }
-    else if (pos >= 0) {
-        return pos;
-    }
-    else if (pos < -(lua_Integer)len) {
-        return 0;
-    }
-    else if (pos < 0) {
-        return len + pos + 1;
-    }
-
-    ASSERT_UNREACHABLE();
-}
-
-static size_t s_grapheme_sub(const char * restrict src, size_t src_len, char * restrict dest, size_t dest_len,
-    size_t start, size_t end)
-{
-    assert(start <= end);
-
-    size_t i = 1;
-
-    // Count graphemes until we reach the start position
-    size_t src_off = 0;
-    while (i < start && src_off < src_len) {
-        size_t inc = grapheme_next_character_break_utf8(src + src_off, src_len - src_off);
-        src_off += inc;
-        i++;
-    }
-
-    // Copy graphemes until we reach the end position (or the buffers end)
-    size_t dest_off = 0;
-    while (i <= end && src_off < src_len && dest_off < dest_len) {
-        size_t inc = grapheme_next_character_break_utf8(src + src_off, src_len - src_off);
-        if (dest != NULL) {
-            memcpy(dest + dest_off, src + src_off, inc);
-        }
-        src_off += inc;
-        dest_off += inc;
-        i++;
-    }
-
-    return dest_off;
-}
-
-static int uni_sub(lua_State *L)
-{
-    // Retrieve the source string and its length
-    size_t byte_len;
-    const char *src = luaL_checklstring(L, 1, &byte_len);
-
-    // Count graphemes. This is necessary to normalize the start and end positions.
-    size_t grapheme_len = c_count_graphemes(src, byte_len);
-
-    // Retrieve the start and end positions
-    size_t start = normalize_start_pos(luaL_checkinteger(L, 2), grapheme_len);
-    size_t end = normalize_end_pos(luaL_optinteger(L, 3, -1), grapheme_len);
-
-    // Bail out if the start offset is greater than the end offset
-    if (start > end) {
-        lua_pushliteral(L, "");
-        return 1;
-    }
-
-    // First pass: determine the length of the substring
-    size_t dest_len = s_grapheme_sub(src, byte_len, NULL, SIZE_MAX, start, end);
-
-    // Prepare the destination buffer
-    luaL_Buffer b;
-    char *dest = luaL_buffinitsize(L, &b, dest_len);
-
-    // Second pass: copy the substring into the destination buffer
-    s_grapheme_sub(src, byte_len, dest, dest_len, start, end);
-
-    // Push the destination string onto the stack
-    luaL_pushresultsize(&b, dest_len);
-
-    return 1;
-}
-
 static luaL_Reg funcs[] = {
     {                  "lower",                  lower},
     {                  "upper",                  upper},
@@ -113,7 +11,7 @@ static luaL_Reg funcs[] = {
     {        "grapheme_breaks",        grapheme_breaks},
     {              "graphemes",              graphemes},
     {                "reverse",      reverse_graphemes},
-    {                    "sub",                uni_sub},
+    {                    "sub",          graphemes_sub},
     {                     NULL,                   NULL}
 };
